changed payload type and tried to fix non working interrupt signal catch

+++ cmd/websocket-server/chat.go
@@ -57,7 +57,7 @@ func newChatServer() *chatServer {
 }
 
 type message struct {
-	index int
+	index uint8
 	state bool
 }
